refactor: extract the build-input digest and its two stamps into their own module
